#29 Refactor validator commands folder
