+++ src/factories/commands/order/OrderCancelCommandBuilder.js
@@ -2,7 +2,7 @@ const logger = require('../../../logger')
 const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
 const OrderCancelCommand = require('../../../commands/order/OrderCancelCommand')
-const OrderCancelCommandValidator = require('../../../validators/OrderCancelCommandValidator')
+const OrderCancelCommandValidator = require('../../../validators/commands/order/OrderCancelCommandValidator')
 const OrderService = require('../../../services/OrderService')
 const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
