+++ src/factories/commands/order/OrderSignCommandBuilder.js
@@ -2,7 +2,7 @@ const logger = require('../../../logger')
 const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
 const OrderSignCommand = require('../../../commands/order/OrderSignCommand')
-const OrderSignCommandValidator = require('../../../validators/OrderSignCommandValidator')
+const OrderSignCommandValidator = require('../../../validators/commands/order/OrderSignCommandValidator')
 const OrderSignerHelper = require('../../../helpers/OrderSignerHelper')
 const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
