+++ src/factories/commands/token/GetAllowanceCommandBuilder.js
@@ -2,7 +2,7 @@ const logger = require('../../../logger')
 const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
 const GetAllowanceCommand = require('../../../commands/token/GetAllowanceCommand')
-const GetAllowanceCommandValidator = require('../../../validators/GetAllowanceCommandValidator')
+const GetAllowanceCommandValidator = require('../../../validators/commands/token/GetAllowanceCommandValidator')
 const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
 
 const privateKeyService = new PrivateKeyService(logger)
