+++ src/factories/commands/trading-wallet/WithdrawCommandBuilder.js
@@ -2,7 +2,7 @@ const logger = require('../../../logger')
 const { PrivateKeyService } = require('../../../services/PrivateKeyService')
 
 const WithdrawCommand = require('../../../commands/trading-wallet/WithdrawCommand')
-const WithdrawCommandValidator = require('../../../validators/WithdrawCommandValidator')
+const WithdrawCommandValidator = require('../../../validators/commands/trading-wallet/WithdrawCommandValidator')
 const PrivateKeyValidator = require('../../../validators/PrivateKeyValidator')
 const TradingWalletServiceBuilder = require('../../TradingWalletServiceBuilder')
