+++ src/validators/commands/trading-wallet/DepositEthCommandValidator.js
@@ -1,5 +1,5 @@
-const customJoiValidator = require('../utils/customJoiValidator')
-const BaseValidator = require('./BaseValidator')
+const customJoiValidator = require('../../../utils/customJoiValidator')
+const BaseValidator = require('../../BaseValidator')
 
 const depositEthSchema = customJoiValidator.object()
   .keys({
