+++ src/validators/commands/trading-wallet/DepositTokenCommandValidator.js
@@ -1,5 +1,5 @@
-const customJoiValidator = require('../utils/customJoiValidator')
-const BaseValidator = require('./BaseValidator')
+const customJoiValidator = require('../../../utils/customJoiValidator')
+const BaseValidator = require('../../BaseValidator')
 
 const depositTokenSchema = customJoiValidator.object()
   .keys({
