+++ src/validators/commands/trading-wallet/GetAddressCommandValidator.js
@@ -1,5 +1,5 @@
-const BaseValidator = require('./BaseValidator')
-const customJoiValidator = require('../utils/customJoiValidator')
+const BaseValidator = require('../../BaseValidator')
+const customJoiValidator = require('../../../utils/customJoiValidator')
 
 const getTradingWalletAddressSchema = customJoiValidator.object()
   .keys({ personalWalletAddress: customJoiValidator.address().ethereum().required(),
