+++ src/validators/index.js
@@ -1,8 +1,8 @@
 const BaseValidator = require('./BaseValidator')
-const CreateWalletCommandValidator = require('./CreateWalletCommandValidator')
-const GetAddressCommandValidator = require('./GetAddressCommandValidator')
+const CreateWalletCommandValidator = require('./commands/trading-wallet/CreateWalletCommandValidator')
+const GetAddressCommandValidator = require('./commands/trading-wallet/GetAddressCommandValidator')
 const OrderPayloadBuilderValidator = require('./OrderPayloadBuilderValidator')
-const OrderSignCommandValidator = require('./OrderSignCommandValidator')
+const OrderSignCommandValidator = require('./commands/order/OrderSignCommandValidator')
 const PrivateKeyValidator = require('./PrivateKeyValidator')
 const TransactionValidator = require('./TransactionValidator')
