+++ tests/unit/commands/createWalletCommand.test.js
@@ -6,7 +6,7 @@ const { EidooEthApiLib } = require('@eidoo/ethapi-lib')
 const { PrivateKeyService } = require('../../../src/services/PrivateKeyService')
 const { TransactionLib } = require('../../../src/lib/TransactionLib')
 const CreateWalletCommand = require('../../../src/commands/trading-wallet/CreateWalletCommand')
-const CreateWalletCommandValidator = require('../../../src/validators/CreateWalletCommandValidator')
+const CreateWalletCommandValidator = require('../../../src/validators/commands/trading-wallet/CreateWalletCommandValidator')
 const logger = require('../../../src/logger')
 const PrivateKeyValidator = require('../../../src/validators/PrivateKeyValidator')
 const TradingWalletService = require('../../../src/services/TradingWalletService')
