+++ tests/unit/commands/getTradingWalletBalanceCommand.test.js
@@ -7,7 +7,7 @@ const logger = require('../../../src/logger')
 const TradingWalletServiceBuilder = require('../../../src/factories/TradingWalletServiceBuilder')
 const PrivateKeyValidator = require('../../../src/validators/PrivateKeyValidator')
 
-const GetBalanceCommandValidator = require('../../../src/validators/GetBalanceCommandValidator')
+const GetBalanceCommandValidator = require('../../../src/validators/commands/trading-wallet/GetBalanceCommandValidator')
 const GetBalanceCommand = require('../../../src/commands/trading-wallet/GetBalanceCommand')
 
 const getBalanceCommandValidator = new GetBalanceCommandValidator(logger)
